+++ .run/publishAll.run.xml
@@ -11,7 +11,6 @@
       <option name="taskNames">
         <list>
           <option value=":thymeleaf:publishToMavenLocal" />
-          <option value=":jte-compiler:publishToMavenLocal" />
           <option value=":jte:publishToMavenLocal" />
           <option value=":core:publishToMavenLocal" />
           <option value=":kte:publishToMavenLocal" />
