+++ core/src/testFixtures/kotlin/de/tschuehly/spring/viewcomponent/core/IntegrationTestBase.kt
@@ -17,14 +17,13 @@ abstract class IntegrationTestBase {
             //language=html
             """
                 <html>
+                    <head><script src="http://localhost:35729/livereload.js"></script></head>
                     <div>
                         <h1>This is the IndexViewComponent</h1>
                     </div>
                     <a href="/">IndexViewComponent</a><br>
                     <a href="/simple">SimpleViewComponent</a><br>
                     <a href="/layout">LayoutViewComponent</a><br>
-                    <a href="/action">ActionViewComponent</a><br>
-                    <a href="/nested-action">NestedActionViewComponent</a><br>
                 </html>
             """.trimIndent()
         assertEndpointReturns("/", expectedHtml)
