+++ examples/jte-example/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.view-component.local-development=false
+spring.view-component.local-development=true
 spring.view-component.view-component-root=examples/jte-example/src/main/java
 server.port=8082
