+++ examples/kte-example/src/main/kotlin/de/tschuehly/kteviewcomponentexample/web/index/IndexViewComponent.kte
@@ -1,4 +1,7 @@
 <html>
+<head>
+    <script src="http://localhost:35729/livereload.js"></script>
+</head>
 <div>
     <h1>This is the IndexViewComponent</h1>
