+++ examples/kte-example/src/main/resources/application.properties
@@ -1,6 +1,3 @@
 server.port=8081
 spring.view-component.local-development=true
-
-gg.jte.templateLocation=examples/kte-example/src/main/kotlin
-gg.jte.templateSuffix=.kte
-gg.jte.developmentMode=true
+spring.view-component.view-component-root=examples/kte-example/src/main/kotlin
