--- examples/kte-example/src/test/resources/application.yaml
@@ -1,9 +1,0 @@
-gg.jte.templateLocation: jte/src/test/kotlin
-logging:
-  level:
-    org.springframework.web: TRACE
-server:
-  port: 8084
-spring:
-  view-component:
-    local-development: false
