+++ examples/thymeleaf-java-example/src/main/java/de/tschuehly/example/thymeleafjava/web/index/IndexViewComponent.html
@@ -1,8 +1,11 @@
 <html>
-    <div>
-        <h1>This is the IndexViewComponent</h1>
-    </div>
-    <a href="/">IndexViewComponent</a><br>
-    <a href="/simple">SimpleViewComponent</a><br>
-    <a href="/layout">LayoutViewComponent</a><br>
+<head>
+  <script src="http://localhost:35729/livereload.js"></script>
+</head>
+<div>
+  <h1>This is the IndexViewComponent</h1>
+</div>
+<a href="/">IndexViewComponent</a><br>
+<a href="/simple">SimpleViewComponent</a><br>
+<a href="/layout">LayoutViewComponent</a><br>
 </html>
