+++ examples/thymeleaf-kotlin-example/src/main/kotlin/de/tschuehly/example/thymeleafkotlin/web/index/IndexViewComponent.html
@@ -1,4 +1,7 @@
 <html>
+<head>
+  <script src="http://localhost:35729/livereload.js"></script>
+</head>
 <div>
   <h1>This is the IndexViewComponent</h1>
