+++ examples/thymeleaf-kotlin-example/src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8084
-spring.view-component.viewComponentRoot=examples/thymeleaf-kotlin-example/src/main/kotlin
+spring.view-component.view-component-root=examples/thymeleaf-kotlin-example/src/main/kotlin
 spring.view-component.local-development=true
