+++ examples/thymeleaf-kotlin-example/src/test/kotlin/de/tschuehly/example/thymeleafkotlin/ThymeleafKotlinIntegrationLocalDevTest.kt
@@ -1,5 +1,4 @@
-package de.tschuehly.kteviewcomponentexample
-
+package de.tschuehly.example.thymeleafkotlin
 
 import de.tschuehly.spring.viewcomponent.core.IntegrationTestBase
 import org.springframework.boot.test.context.SpringBootTest
@@ -7,6 +6,6 @@ import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest(
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    properties = ["spring.view-component.local-development=false"]
+    properties = ["spring.view-component.local-development=true"]
 )
-class KteIntegrationTest : IntegrationTestBase()
+class ThymeleafKotlinIntegrationLocalDevTest : IntegrationTestBase()
