+++ examples/thymeleaf-kotlin-example/src/test/kotlin/de/tschuehly/example/thymeleafkotlin/ThymeleafKotlinIntegrationTest.kt
@@ -5,6 +5,7 @@ import org.springframework.boot.test.context.SpringBootTest
 
 
 @SpringBootTest(
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
+    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+    properties = ["spring.view-component.local-development=false"]
 )
 class ThymeleafKotlinIntegrationTest : IntegrationTestBase()
