+++ examples/thymeleaf-kotlin-example/src/test/resources/application.properties
@@ -1,2 +1,2 @@
-spring.view-component.local-development=false
 server.port=8084
+spring.view-component.view-component-root=src/main/kotlin
