--- kte/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-gg.jte.templateLocation=src/main/kotlin
-gg.jte.templateSuffix=.kte
