+++ kte/src/main/resources/kte.properties
@@ -1,0 +1,2 @@
+spring.view-component.view-component-root=src/main/kotlin
+spring.view-component.template-suffix=.kte
